requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib
python-jose
redis
azure-storage-blob